+++ scripts/build.sh
@@ -5,7 +5,7 @@ export DIP_DELIVERABLE="${DIP_DELIVERABLE:-dip}"
 
 echo "GITHUB_TAG: '$GITHUB_TAG' DIP_DELIVERABLE: '${DIP_DELIVERABLE}'"
 cd cmd/dip
-go build -ldflags "-X main.Version=${GITHUB_TAG}" -o "${DIP_DELIVERABLE}"
+go build -buildvcs=false -ldflags "-X main.Version=${GITHUB_TAG}" -o "${DIP_DELIVERABLE}"
 $SHA512_CMD "${DIP_DELIVERABLE}" > "${DIP_DELIVERABLE}.sha512.txt"
 chmod +x "${DIP_DELIVERABLE}"
 cd ../..